+++ src/Microsoft.Data.SqlClient/netcore/src/Microsoft/Data/SqlClient/Server/MetadataUtilsSmi.cs
@@ -391,8 +391,8 @@ namespace Microsoft.Data.SqlClient.Server
             return s_extendedTypeCodeToSqlDbTypeMap[(int)typeCode + 1];
         }
 
-        // Infer SqlDbType from Type in the general case.  Katmai-only (or later) features that need to 
-        //  infer types should use InferSqlDbTypeFromType_Katmai.
+        // Infer SqlDbType from Type in the general case. 2008-only (or later) features that need to 
+        //  infer types should use InferSqlDbTypeFromType_2008.
         internal static SqlDbType InferSqlDbTypeFromType(Type type)
         {
             ExtendedClrTypeCode typeCode = DetermineExtendedTypeCodeFromType(type);
@@ -409,12 +409,12 @@ namespace Microsoft.Data.SqlClient.Server
             return returnType;
         }
 
-        // Inference rules changed for Katmai-or-later-only cases.  Only features that are guaranteed to be 
-        //  running against Katmai and don't have backward compat issues should call this code path.
-        //      example: TVP's are a new Katmai feature (no back compat issues) so can infer DATETIME2
+        // Inference rules changed for 2008-or-later-only cases.  Only features that are guaranteed to be 
+        //  running against 2008 and don't have backward compat issues should call this code path.
+        //      example: TVP's are a new 2008 feature (no back compat issues) so can infer DATETIME2
         //          when mapping System.DateTime from DateTable or DbDataReader.  DATETIME2 is better because
         //          of greater range that can handle all DateTime values.
-        internal static SqlDbType InferSqlDbTypeFromType_Katmai(Type type)
+        internal static SqlDbType InferSqlDbTypeFromType_2008(Type type)
         {
             SqlDbType returnType = InferSqlDbTypeFromType(type);
             if (SqlDbType.DateTime == returnType)
@@ -533,7 +533,7 @@ namespace Microsoft.Data.SqlClient.Server
         // Extract metadata for a single DataColumn
         internal static SmiExtendedMetaData SmiMetaDataFromDataColumn(DataColumn column, DataTable parent)
         {
-            SqlDbType dbType = InferSqlDbTypeFromType_Katmai(column.DataType);
+            SqlDbType dbType = InferSqlDbTypeFromType_2008(column.DataType);
             if (InvalidSqlDbType == dbType)
             {
                 throw SQL.UnsupportedColumnTypeForSqlProvider(column.ColumnName, column.DataType.Name);
@@ -733,7 +733,7 @@ namespace Microsoft.Data.SqlClient.Server
             }
 
             Type colType = (Type)temp;
-            SqlDbType colDbType = InferSqlDbTypeFromType_Katmai(colType);
+            SqlDbType colDbType = InferSqlDbTypeFromType_2008(colType);
             if (InvalidSqlDbType == colDbType)
             {
                 // Unknown through standard mapping, use VarBinary for columns that are Object typed, otherwise error